Move Phpseclib adapter into its own namespace
