+++ src/Gaufrette/Adapter/Phpseclib/.travis.yml
@@ -1,0 +1,23 @@
+language: php
+
+sudo: false
+
+cache:
+  directories:
+    - $HOME/.composer/cache/files
+
+php:
+  - 5.6
+  - 7.0
+  - hhvm
+
+matrix:
+  allow_failures:
+    - php: hhvm
+
+before_script:
+  - composer install --prefer-source --no-interaction
+
+script:
+  - vendor/bin/phpspec run -fpretty --verbose --no-interaction
+  - vendor/bin/phpunit
