+++ src/Gaufrette/Adapter/Phpseclib/phpspec.yml.dist
@@ -1,0 +1,5 @@
+suites:
+  main:
+    namespace: Gaufrette\Adapter\Phpseclib\
+    psr4_prefix: Gaufrette\Adapter\Phpseclib\
+    src_path: src/
