+++ src/Gaufrette/Adapter/Phpseclib/phpunit.xml.dist
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<phpunit backupGlobals="false"
+         backupStaticAttributes="false"
+         colors="true"
+         convertErrorsToExceptions="true"
+         convertNoticesToExceptions="true"
+         convertWarningsToExceptions="true"
+         processIsolation="false"
+         stopOnFailure="false"
+         syntaxCheck="false"
+         bootstrap="vendor/autoload.php"
+>
+    <php>
+        <!-- <env name="SFTP_HOST" value="localhost" />
+        <env name="SFTP_PORT" value="22" />
+        <env name="SFTP_USER" value="gaufrette" />
+        <env name="SFTP_PASS" value="gaufrette" />
+        <env name="SFTP_BASE_DIR" value="/home/gaufrette/share" /> -->
+    </php>
+
+    <testsuites>
+        <testsuite name="Phpseclib Adapter Test Suite">
+            <directory>./tests</directory>
+        </testsuite>
+    </testsuites>
+    <filter>
+        <whitelist>
+            <directory>./src</directory>
+        </whitelist>
+    </filter>
+</phpunit>
