+++ tests/Gaufrette/Functional/FileStream/filesystem/copy1.txt
@@ -1,0 +1,1 @@
+test content
