+++ tests/Gaufrette/Functional/FileStream/filesystem/copy2.txt
@@ -1,0 +1,1 @@
+test content
